requirements.txt
ccxt
pandas
sqlalchemy
matplotlib
numpy
backtrader
telegram
python-dotenv